set sweet alert form all legal action

+++ src/Pages/MyCart.jsx
@@ -1,23 +1,36 @@
 import { Link, useLoaderData } from "react-router-dom";
 import Navber from "../Components/Navber";
 import { useState } from "react";
+import Swal from "sweetalert2";
 
 const MyCart = () => {
   const loadedproducts = useLoaderData();
   const [products, setProducts] = useState(loadedproducts);
   const handleDelete = (_id) => {
-    fetch(`http://localhost:5000/products/${_id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        if (data.deletedCount < 0) {
-          alert("data deleted successfully");
-        }
-        const remaining = products.filter((product) => product._id !== _id);
-        setProducts(remaining);
-      });
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/products/${_id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            console.log(data);
+            if (data.deletedCount < 0) {
+              Swal.fire("Deleted!", "Your file has been deleted.", "success");
+            }
+            const remaining = products.filter((product) => product._id !== _id);
+            setProducts(remaining);
+          });
+      }
+    });
   };
   return (
     <div>
